No disclaimer option. Refactoring.
